# api/openapi.yaml
openapi: 3.0.3
info:
  title: TraceLayer Go API
  version: 0.1.0
  description: Current public HTTP contract for the TraceLayer backend.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/v1/correlate:
    post:
      operationId: correlate
      responses:
        '200':
          description: Correlation, entity resolution, and ranking result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrelateResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  schemas:
...
    CorrelateResponse:
      type: object
      required: [status, entities_clustered, leads_generated, observations_correlated]
      properties:
        status: {type: string}
        entities_clustered: {type: integer}
        leads_generated: {type: integer}
        observations_correlated: {type: integer}
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, field]
          properties:
            code: {type: string}
            message: {type: string}
            field: {type: string, nullable: true}
  responses:
...
    InternalError:
      description: Internal or persistence failure
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}